additional solutions

+++ Heart_rate.cpp
@@ -6,9 +6,9 @@ using namespace std;
 int main(){
     int n= 0;
     int b= 0;
-    int p= 0;
-    int diff= 0;
-    int BPM= 0, min_BPM= 0, max_BPM= 0;
+    double p= 0;
+    double diff= 0;
+    double BPM= 0, min_BPM= 0, max_BPM= 0;
     cin >> n;
     for(int i= 0; i< n; i++){
         cin >> b;
